+++ backend/plugins/argo/tasks/deployment_collector.go
@@ -25,7 +25,7 @@ import (
 	"github.com/apache/incubator-devlake/core/errors"
 	"github.com/apache/incubator-devlake/core/plugin"
 	helper "github.com/apache/incubator-devlake/helpers/pluginhelper/api"
-	"github.com/apache/incubator-devlake/plugins/ra_dora/models"
+	"github.com/apache/incubator-devlake/plugins/argo/models"
 )
 
 const RAW_DEPLOYMENT_TABLE = "argo_api_deployments"
