+++ backend/plugins/argo/tasks/deployment_convertor.go
@@ -23,7 +23,7 @@ import (
 	"github.com/apache/incubator-devlake/core/errors"
 	"github.com/apache/incubator-devlake/core/plugin"
 	"github.com/apache/incubator-devlake/helpers/pluginhelper/api"
-	"github.com/apache/incubator-devlake/plugins/ra_dora/models"
+	"github.com/apache/incubator-devlake/plugins/argo/models"
 )
 
 var _ plugin.SubTaskEntryPoint = ConvertDeployments
@@ -40,7 +40,7 @@ var ConvertDeploymentsMeta = plugin.SubTaskMeta{
 //
 //	"github.com/apache/incubator-devlake/core/models/domainlayer/devops"
 func ConvertDeployments(taskCtx plugin.SubTaskContext) errors.Error {
-	log.Println("Iniciando plugin de convert.")
+	log.Println("[ARGO] Iniciando plugin de convert.")
 
 	// TODO
 	converter, err := api.NewStatefulDataConverter(&api.StatefulDataConverterArgs[models.Deployment]{
@@ -56,6 +56,6 @@ func ConvertDeployments(taskCtx plugin.SubTaskContext) errors.Error {
 		return err
 	}
 
-	log.Println("Conversão de deployments concluída com sucesso!")
+	log.Println("[ARGO] Conversão de deployments concluída com sucesso!")
 	return converter.Execute()
 }
