+++ backend/plugins/argo/tasks/deployment_extractor.go
@@ -25,7 +25,7 @@ import (
 	"github.com/apache/incubator-devlake/core/models/common"
 	"github.com/apache/incubator-devlake/core/plugin"
 	"github.com/apache/incubator-devlake/helpers/pluginhelper/api"
-	"github.com/apache/incubator-devlake/plugins/ra_dora/models"
+	"github.com/apache/incubator-devlake/plugins/argo/models"
 )
 
 var _ plugin.SubTaskEntryPoint = ExtractDeployments
